KupData: moved "applyFormulaToColumns" to "datasetOperations.column".
