+++ packages/ketchup/src/managers/kup-data/kup-data-column-helper.ts
@@ -4,6 +4,11 @@ import {
     DataTable,
     Row,
 } from '../../components/kup-data-table/kup-data-table-declarations';
+import { KupDebugCategory } from '../kup-debug/kup-debug-declarations';
+import { KupLanguageTotals } from '../kup-language/kup-language-declarations';
+import { KupDom } from '../kup-manager/kup-manager-declarations';
+
+const dom: KupDom = document.documentElement as KupDom;
 
 /**
  * Finds the columns matching the criteria specified in the filters argument.
@@ -95,5 +100,141 @@ export function mergeColumns(
         rowIndex++;
     }
     dataset.columns.push(newColumn);
+    return newColumn;
+}
+/**
+ * This method is used to create a new column from a mathematical formula.
+ * @param {DataTable} dataset - The dataset that must be updated with the new columns.
+ * @param {string} operation - Mathematical operation to apply (i.e.: "sum", "average", ([COL1] - [COL2]) * 100 / [COL3]).
+ * @param {string[]} columns - Column names used for the mathematical operation. When missing, they will be extracted from the formula.
+ * @returns {string | Column} Returns the new column created or a string containing the error message (if something went wrong).
+ */
+export function newColumn(
+    dataset: DataTable,
+    operation: string,
+    columns?: string[]
+): string | Column {
+    if (!columns) {
+        columns = [];
+    }
+    if (columns.length === 0) {
+        const names = operation.split('[');
+        for (let i = 1; i < names.length; i++) {
+            columns.push(names[i].split(']')[0]);
+        }
+    }
+    if (columns.length === 0) {
+        const message =
+            "Can't apply math formulas without columns!(" + columns + ')';
+        dom.ketchup.debug.logMessage(
+            'kup-data',
+            message,
+            KupDebugCategory.WARNING
+        );
+        return message;
+    }
+    const titles: string[] = [];
+    const formulaRow: { [index: string]: number } = {};
+    let firstColumn: Column = null;
+    let formula = '';
+    switch (operation) {
+        case KupLanguageTotals.AVERAGE:
+            formula = `(${columns.join(' + ')}) / ${columns.length}`;
+            break;
+        case KupLanguageTotals.DIFFERENCE:
+            formula = columns.join(' - ');
+            break;
+        case KupLanguageTotals.PRODUCT:
+            formula = columns.join(' * ');
+            break;
+        case KupLanguageTotals.SUM:
+            formula = columns.join(' + ');
+            break;
+        default:
+            formula = operation;
+    }
+    for (let index = 0; index < dataset.columns.length; index++) {
+        const col = dataset.columns[index];
+        if (columns.includes(col.name)) {
+            titles[columns.indexOf(col.name)] = col.title;
+            if (!dom.ketchup.objects.isNumber(col.obj)) {
+                const message =
+                    "Can't apply math formulas on non-numerical columns!(" +
+                    columns +
+                    ')';
+                dom.ketchup.debug.logMessage(
+                    'kup-data',
+                    message,
+                    KupDebugCategory.WARNING
+                );
+                return message;
+            }
+        }
+        if (columns[0] === col.name) {
+            firstColumn = col;
+        }
+        if (col.resultOf && col.resultOf === formula) {
+            const message =
+                'This mathematical operation on these columns was already performed!(' +
+                formula +
+                ')';
+            dom.ketchup.debug.logMessage(
+                'kup-data',
+                message,
+                KupDebugCategory.WARNING
+            );
+            return message;
+        }
+    }
+    let prog = 0;
+    let newName = 'MATH_';
+    while (findColumns(dataset, { name: newName + prog }).length > 0) {
+        prog++;
+    }
+    newName = newName + prog;
+    const newObj = firstColumn.obj;
+    let newTitle = formula;
+    for (let i = 0; i < columns.length; i++) {
+        const column = columns[i];
+        let re: RegExp = new RegExp(column, 'g');
+        newTitle = newTitle.replace(re, titles[i]);
+    }
+    dataset.rows.forEach((row) => {
+        const cells = row.cells;
+        let base: Cell = null;
+        if (cells) {
+            for (let index = 0; index < columns.length; index++) {
+                const column = columns[index];
+                const cell = cells[column];
+                if (cell) {
+                    if (!base) {
+                        base = cell;
+                    }
+                    formulaRow[column] = dom.ketchup.data.numberify(cell.value);
+                }
+            }
+        }
+        const value = dom.ketchup.data.formulas
+            .custom(formula, formulaRow)
+            .toString();
+        cells[newName] = {
+            ...base,
+            displayedValue: null,
+            obj: { ...newObj, k: value },
+            value: value,
+        };
+    });
+    const newColumn: Column = {
+        ...firstColumn,
+        name: newName,
+        title: newTitle,
+        obj: newObj,
+        resultOf: formula,
+    };
+    dataset.columns.splice(
+        dataset.columns.indexOf(firstColumn) + 1,
+        0,
+        newColumn
+    );
     return newColumn;
 }
